Fixed implementation of EventHandler by using Uintptr directly and finishing events example

+++ src/WebUI.NET/Events/EventType.cs
@@ -8,7 +8,7 @@
 
 namespace WebUI.Events
 {
-    public enum EventType : uint
+    public enum EventType : ulong
     {
         Disconnect = 0,
         Connect = 1,
